- add renew log  - add evit log

+++ server/server/meta/src/main/java/com/alipay/sofa/registry/server/meta/lease/session/DefaultSessionServerManager.java
@@ -117,12 +117,12 @@ public class DefaultSessionServerManager
   }
 
   @Override
-  protected long getIntervalMilli() {
+  protected int getIntervalMilli() {
     return metaServerConfig.getExpireCheckIntervalMillis();
   }
 
   @Override
-  protected long getEvictBetweenMilli() {
+  protected int getEvictBetweenMilli() {
     return metaServerConfig.getExpireCheckIntervalMillis();
   }
